pyproject.toml
[tool.poetry]
name = "mypackage"
version = "0.0.1"
description = "A package to demo how the Hydra framework can be used with SageMaker"
authors = ["David Sauerwein <dsauerwe@amazon.com>", "Julian Grueber <juligrue@amazon.com>", "Marco Geiger <marcogei©amazon.ch"]
license = "MIT No Attribution"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
hydra-core = "^1.3.2"
boto3 = "^1.34.39"
numpy = "^1.26.4"
pandas = "^2.2.0"
scikit-learn = "^1.4.0"
matplotlib = "^3.8.2"

[tool.poetry.group.dev.dependencies]
commitizen = "^3.12.0"
bandit = "^1.7.5"
pre-commit = "^3.5.0"
pydocstyle = "^6.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
